refactor: extract avatar component
