+++ src/lib/AboutMe.svelte
@@ -1,31 +1,14 @@
 <script lang="ts">
+	import FlippingAvatar from './FlippingAvatar.svelte';
+
 	import { InstagramLogo } from 'radix-icons-svelte';
 	import { QRCode, ScrollDownButton } from '$lib';
 
 	export let portrait: string;
 	export let portfolio: string;
 
-	let isPortraitFlipping: boolean = false;
-	let isCardOnPortraitSide: boolean = true;
-
-	function handleAvatarClick(event: MouseEvent & { currentTarget: EventTarget & HTMLElement }) {
-		isPortraitFlipping = true;
-		console.log(event);
-	}
-
-	function handleAvatarAnimationEnd(
-		event: AnimationEvent & { currentTarget: EventTarget & HTMLElement }
-	) {
-		isPortraitFlipping = false;
-		isCardOnPortraitSide = !isCardOnPortraitSide;
-	}
-
 	const bioDescription =
 		'Innovative fashion designer and stylist, graduate of the Artistic School of Fashion Design,	with more than 3 years experience in the fashion industry. Proficient in conceptualizing and executing fashion shoots and managing styling for various editorial projects. Experienced in working with models, photographers, and creative teams to create captivating and high-quality images. Ready to bring fresh, trendsetting ideas to your brand.';
-
-	$: {
-		console.log(isCardOnPortraitSide, isPortraitFlipping);
-	}
 </script>
 
 <section
@@ -70,27 +53,7 @@
 		</div>
 		<div
 			class="h-[70svh] lg:h-[80svh] portrait:h-[50svh] w-full p-2 sm:p-4 lg:p-12 row-start-1 landscape:col-start-3">
-			<!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
-			<!-- svelte-ignore a11y-click-events-have-key-events -->
-			<div class="h-full w-full">
-				<img
-					class="h-full w-full object-scale-down portrait:max-md:backface-hidden {isPortraitFlipping &&
-						(isCardOnPortraitSide
-							? 'animate-flip-y'
-							: 'animate-flip-y-reverse')} {isCardOnPortraitSide ? 'rotate-y-0' : 'rotate-y-180'}"
-					src={portrait}
-					alt="Portrait of a beautiful girl."
-					on:click={handleAvatarClick}
-					on:animationend={handleAvatarAnimationEnd} />
-				<p
-					class="hidden portrait:max-md:block h-full overflow-y-auto overflow-x-hidden w-full px-6 text-justify text-lg relative -top-full left-[50%] translate-x-[-50%] backface-hidden {isPortraitFlipping &&
-						(isCardOnPortraitSide
-							? 'animate-flip-y-reverse'
-							: 'animate-flip-y')} {isCardOnPortraitSide ? 'rotate-y-180' : 'rotate-y-0'}"
-					on:click={handleAvatarClick}>
-					{bioDescription}
-				</p>
-			</div>
+			<FlippingAvatar {portrait} {bioDescription} />
 		</div>
 	</div>
 </section>
